Update
Change MONITOR_TIME from every 30 seconds to every 40 seconds. Remember max is 45 before BattlEye disconnects you.

Pushed Socket Exception up to creation of JconClient.

Removed uneeded InterruptedException, and try catch statements.

Command auth before sending packet, so non-commands are not sent to rcon to return "Unknown Command".

Added a debugging System.exit to stop the program whenever the dayz server crashes, or is closed. Used for debugging.

Currently, I am at around 2.5 hours of access to rcon before any issues occurs; however, still that means that there is a chance for the library to crash your BattlEye protected server.
I have no been able to identify why this is the case as pushing the fequency of the library which it is sending null command packets as not seemed to introduce, or fixed performance issues.